--- packages/cli/test/cases/build.default.workspace-template-page-style/src/styles/theme.css
@@ -1,15 +1,0 @@
-@import url('//fonts.googleapis.com/css?family=Source+Sans+Pro&display=swap');
-
-* {
-  margin: 0;
-  padding: 0;
-  text-decoration: none;
-}
-
-html {
-  background-color: #f0f0f0;
-}
-
-body {
-  font-family: 'Source Sans Pro', sans-serif;
-}
