+++ packages/cli/test/cases/build.default.workspace-template-page/build.default.workspace-template-page.spec.js
@@ -22,6 +22,7 @@
  */
 const expect = require('chai').expect;
 const fs = require('fs');
+const glob = require('glob-promise');
 const { JSDOM } = require('jsdom');
 const path = require('path');
 const TestBed = require('../../../../../test/test-bed');
@@ -53,12 +54,6 @@ describe('Build Greenwood With: ', function() {
         expect(fs.existsSync(path.join(this.context.publicDir, './index.html'))).to.be.true;
       });
 
-      it('should have the specific element we added as part of our custom page template', function() {
-        const customElement = dom.window.document.querySelectorAll('div.owen-test');
-
-        expect(customElement.length).to.equal(1);
-      });
-
       describe('correct merge order for default app and custom page template <head> tags', function() {
         let scriptTags;
         let linkTags;
@@ -87,12 +82,62 @@ describe('Build Greenwood With: ', function() {
         });
 
         it('should add one page template <link> tag', function() {
-          expect(linkTags[0].href).to.match(/styles\/theme.*.css/);
+          expect(linkTags[0].rel).to.equal('stylesheet');
+          expect(linkTags[0].href).to.match(/styles\/theme.[a-z0-9]{8}.css/);
         });
 
         it('should add one page template <style> tag', function() {
           // offset index by one since first <style> tag is from Puppeteer
           expect(styleTags[1].textContent).to.contain('.owen-test');
+        });
+      });
+
+      describe('custom <link> tag in the <head> of a page template', function() {
+        let dom;
+
+        before(async function() {
+          dom = await JSDOM.fromFile(path.resolve(this.context.publicDir, 'index.html'));
+        });
+  
+        it('should generate a single matching CSS file with hashed filename', async function() {
+          expect(await glob.promise(`${path.join(this.context.publicDir, 'styles')}/theme.*.css`)).to.have.lengthOf(1);
+        });
+  
+        it('should have the expected <link> tag in the <head>', async function() {
+          const linkTags = dom.window.document.querySelectorAll('head > link');
+          const linkTag = linkTags[0];
+  
+          expect(linkTags.length).to.equal(1);
+          expect(linkTag.rel).to.equal('stylesheet');
+          expect((/styles\/theme.[a-z0-9]{8}.css/).test(linkTag.href)).to.be.true;
+        });
+      });
+  
+      describe('custom <style> tag in the <head> of a page template', function() {
+        let dom;
+  
+        before(async function() {
+          dom = await JSDOM.fromFile(path.resolve(this.context.publicDir, 'index.html'));
+        });
+
+        it('should have the specific element we added as part of our custom page template', function() {
+          const customElement = dom.window.document.querySelectorAll('div.owen-test');
+  
+          expect(customElement.length).to.equal(1);
+        });
+  
+        it('should have the color style for the .owen-test element in the page template that we added as part of our custom style', function() {
+          const customElement = dom.window.document.querySelector('.owen-test');
+          const computedStyle = dom.window.getComputedStyle(customElement);
+  
+          expect(computedStyle.color).to.equal('blue');
+        });
+  
+        it('should have the color styles for the h3 element that we defined as part of our custom style', function() {
+          const customHeader = dom.window.document.querySelector('h3');
+          const computedStyle = dom.window.getComputedStyle(customHeader);
+  
+          expect(computedStyle.color).to.equal('green');
         });
       });
     });
