+++ packages/cli/test/cases/build.default.workspace-template-page/src/templates/page.html
@@ -20,13 +20,12 @@
   </head>
 
   <body>
-
+    
     <div class='wrapper'>
-      
-      <div class='page-template blog-content content owen-test'>
+      <div class='page-template content owen-test'>
+        <h3>Hello</h3>
         <content-outlet></content-outlet>
       </div>
-    
     </div>
 
   </body>
